% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeCourseVectors.R
\name{TimeCourseVectors}
\alias{TimeCourseVectors}
\title{TimeCourseVectors}
\description{
An R6 class to transform experimental matrix data into timecourse vectors.
}
\examples{
  tcv <- TimeCourseVectors$new(id="abc")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeCourseVectors-new}{\code{TimeCourseVectors$new()}}
\item \href{#method-TimeCourseVectors-getScoreColumn}{\code{TimeCourseVectors$getScoreColumn()}}
\item \href{#method-TimeCourseVectors-clone}{\code{TimeCourseVectors$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeCourseVectors-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeCourseVectors-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeCourseVectors$new(tbl, scoreColumn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl}}{data.frame}

\item{\code{scoreColumn}}{character}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance of TimeCourseVectors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeCourseVectors-getScoreColumn"></a>}}
\if{latex}{\out{\hypertarget{method-TimeCourseVectors-getScoreColumn}{}}}
\subsection{Method \code{getScoreColumn()}}{
accessor for the object's id field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeCourseVectors$getScoreColumn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the current value of the id member
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeCourseVectors-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeCourseVectors-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeCourseVectors$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
